Disable backported methods for API levels higher than known Android release
If the min API level is set higher than the last known Android
release, the methods implemented by the platform is not known, and the
compiler cannot know which methods to backport. The backporting of
methods is then turned off and a warning is issued.

For the "magic" API level of 10000 the warning is not issued, as that
is the API level used when building the core library in the Android
Platform. Here disabling of backported methods is the expected
behaviour.

Bug: 147480264
Change-Id: I59b96fc0543a243861c7a0e3b0b9428504e96555