Make sure to have an IR instruction materialized for local ends.
nop and move instructions did not materialize in the IR even if
the instructions end the live range of a local. This caused the
IR to sometimes contain an instruction that simultaneously starts
and ends a local. This complicates the rest of the pipeline.

With this change, all instructions that end the live range of
a local materialize either as an actual instruction or as
a DebugLocalRead.

BUG=147865212
R=zerny@google.com

Change-Id: Iedd0190092d6114726ece7fda137fec7bb14e54d