[Retrace] Retracing single line mappings for ranges
When retracing mappings on the form 1:10:void qux():27:27 -> a, where
the obfuscated range points to a single line, we should report the
single line and not compute the offset. This is compatible with the
reporting that PG retrace has.

Also, this will allow us to greatly reduce the size of map if we
produced mappings based on PC, since we can merge a sequence of PC's
that point to the same line.

Bug: 148035703
Change-Id: Id71be8aa9229988356bb0d5e22be979a50982601