Rename callsite different from the implementing interfaces' methods
This is necessary when having default implementations, otherwise the
runtime behavior would lead to un-expected behavior:

  public interface I {
    void foo();
  }

  public interface J {
    default void bar() {
      System.out.println("J.bar");
    }
  }

 public static void main(String[] args) {
   I i = (I & J) () -> System.out.println("Lambda.foo");
   ((J)i).bar();
 }

Running the program above, one would expected it to print J.bar,
however, if we give the same minified name to I.foo and J.bar, the
resolution will find the definition in the callsite, and print
Lambda.foo. This CL tracks all multiple interfaces with methods that
would collide with the bootstrap method and ensures they are renamed
differently.

Bug: 148368442
Change-Id: I7d853afdce393b5a207a5c98639317b1a085ac2c