Compute if LookupResult is incomplete based on keep and library
When querying for dispatch targets, we may not know all the methods
implemented based on library and keep rules. This CL adds multiple
tests for when to mark the result as complete and incomplete and
alters the lookup algorithm to satisfy the tests.

This is not complete yet since we do not still account for call sites

Bug: 149941117
Change-Id: I28ca057baf3badad31695bfdcf6848132adb2489