Towards synthesizing Kotlin @Metadata: type projection (type arguments).
Type parameters in generic types are converted to KmTypeProjection in
Kotlin @Metadata. We can store them in the beginning and rewrite them
accordingly at the end, but that's not a desired solution: it persists
from end to end. Besides, that's not the only source of truth.

Instead, we can retrieve type parameters from Signature attributes.
The only caveat is that we have String-to-String signature rewriter,
not internal encodings. To nail down to specific type parameters for
specific arguments, we really need our own encodings that conform to
syntax of JVM Signature attributes.

Based on the chained CL, where such encodings are added, this CL
add support for synthesizing KmTypeProjection inside KmType.

Bug: 70169921, 143687784
Change-Id: Iccc38f5242097ead253b372ece72c00c1b3bec42