Rewrite lookupSingleVirtual/InterfaceTarget to use lookup
This CL unifies our lookup targets to be defind only on the
resolution. Currently, the implementation here tries to mimic the old
behavior as much as possible when it was not directly incorrect, but
we need to follow up with proper support for looking a library methods
and updates to what is considered incomplete.

Bug: 149941117
Bug: 140204899
Bug: 148429150
Bug: 148584615
Bug: 148591377
Bug: 149363086
Bug: 150640456
Change-Id: Ia71aee3f309cd223d39b21fe83790100237c01e2