Determine the use of not mapped types lazily for applymapping
When compiling an app with already minified input, like the play core
module, R8 will remove some classes and rename others with a high
probability of us rotating names. When running tests afterwards, we
would report an error if we found a mapping to a class that has no
mapping since such types could collide if both were used in the tests.

This CL pushes this tests to the NamingLens such that unused
references in the test will not throw an error. Also, the error is
lowered to a warning.

Bug: 149946708
Change-Id: I70e690532bd512fba870f82e1561955c85519e11