Account for host-companion relations when synthesizing property @Metadata.
Backing fields for properties in companion classes are declared in host
classes. @Metadata synthesizer currently searched for live members of
Kotlin properties _within_ a class, resulting in removals of properties
whose backing fields are not in the companion, but in the host class.

To address that, this CL
1) stores the relations between host and companion classes;
2) restores/rewrites companion object and class in the host class; and
3) revises property builder logic to explore proper fields with proper
tester: for companion, iterates fields in the host class with a tester
that checks if the field is a backing field for companion. Otherwise,
i.e., if we don't distinguish backing field for companion v.s. for the
current class, such field will be grouped to host class's property.

Also, this CL addresses undiscovered issues regarding nested class and
property flags. In particular, property has three different flags---
itself, getter, and setter---and they should be handled differently.
Currently, if a backing field is not needed, getter flags were used
as property flags, resulting in visibility issues.

Bug: 70169921, 143687784
Change-Id: I927dee3ea5dfd4d6e85ecab6b0b509f1f8bde273