Introduce a new InitClass instruction
This CL introduces a new InitClass instruction that can be used to trigger the class initialization of a given type.

The instruction can only be used in R8 when a second round of tree shaking will be run. This is because the Enqueuer is used to collect the set of classes that are referenced from an InitClass instruction. For each of these classes, a new static synthetic field $r8$clinit is synthesized by a new class ClassInitFieldSynthesizer, unless the given class already has another static field that can be used.

The ClassInitFieldSynthesizer produces a InitClassLens that maps each class with an InitClass instruction to the DexField that should be used to trigger class initialization. This is used in application writer, and in particularly, in DexInitClass for implemeting collectIndexedItems() and write().

It is left for future work to add support for InitClass in the CF backend.

Change-Id: I86702e16c42d5190d9dcec1a1fd3150affe9a89b