Don't inline catch-handlers with potentially unknown exception guards.
This CL adds a static list of exceptions introduced at each API level
from 1 up to, but excluding, 21. The list can be generated by a
api-versions.xml file and the generation utility is added for future
reference. The fix for the isssue generalizes the previous fix for the
special case of ReflectiveOperationException.

Bug: 153514654
Change-Id: I983a10b859cdfc995709f49a38763278dc59ed04