Model kotlin metadata in a structured way
This is a major rewrite of kotlin metadata rewriting compared to the
current approach that relied on types, generic signatures and existing
information. Rather, the approach here is straight forward in the
sense that:

- We translate all kotlin metadata structures to an internal
  representation
- When a piece of meta-data is connected with a class or member, we
  attach all needed information to that member
- When rewriting, we simply map over the structure with a naming lense.

This approach is really simply (all though it requires a lot of boiler
plate to get the mapping correct). Additionally, it should make it
easier for tracing structures in the enqueuer.

Bug: 154343420
Bug: 154199572
Bug: 154201250
Bug: 154294232
Bug: 154300326
Bug: 154346948
Bug: 154347404
Bug: 154348149
Bug: 154348568
Bug: 154348683
Bug: 154351125
Bug: 152153136
Change-Id: I92742f04493b28c5e4adcca837ce8202f4d0dab8