Introduce a FieldResolutionResult
This CL changes the field resolution methods on AppInfo to return a FieldResolutionResult to ensure that the initial resolution holder and the resolved field holder is not thrown away. The new FieldResolutionResult defines an isAccessibleFrom() method that can be used to determine if the resolved field is accessible from a given program context. All field accessibility checks in the compiler should be refactored to use this method. This is planned in a follow up.

The enqueuer is refactored to use the new resolution result, to avoid redundant calls to definitionFor().

The majority of the remaining calls to resolveField() are left as-is, except for an additional call to getResolvedField() that retrieves the DexEncodedField (if any). Over time these call sites should be migrated to use the new resolution result.

Bug: 145723539
Change-Id: Ieece1edb2cae685c6c9dea0783a789cb89614b85