Only process and keep kotlin.Metadata when kept
This CL changes the behavior for kotlin metadata to be the same as for
other annotations, that is, one has to keep kotlin.Metadata and also
keep the class for where kotlin metadata should be kept.

If the kotlin metadata describes a lambda, we still process it for us
to be able to lambda merge.

Bug: 154294232
Change-Id: Ie0e3b08c5c90535b25311d6d6dad7917ebc84b46