Introduce an (unused) abandoned call site optimization instance
Abandoned will be used to signal that the call site optimization has bailed out for a method and all of it overrides, such that there is no need to propagate any arguments to the method or its overrides.

If the resolution target of an invoke is abandoned, then we can avoid computing the set of possible dispatch targets in the first place, which is generally expensive.

This also removes two redundant singleton instances of BOTTOM and TOP in the call site optimization.

Change-Id: I7757cc8e2872c6338c325349a56c28e1e7c69c65