Abandon call site optimization when number of targets exceed threshold
When the number of dispatch targets for a given invoke exceed a given threshold, we will now mark all of those targets as 'abandoned'. This allows us to skip the propagation of arguments (including the computation of dispatch targets) if an invoke resolves to a method that is marked as 'abandoned'.

In addition to the above, this CL also makes a few minor optimizations, such that we now bail out as early as possible (e.g., if there are no arguments to propagate) and no longer compute a fresh CallSiteOptimizationInfo object per dispatch target.

Bug: 150269949
Change-Id: I530f47a527a61427cfbaae02d81311c08f809f15