Towards unifying AssumeDynamicType and AssumeNotNull
This extends each Assume instruction to have both a DynamicTypeAssumption and NonNullAssumption.

It is still the case that each Assume instruction in the IR represents an AssumeDynamicType instruction or an AssumeNonNull instruction. A follow up should remove redundant Assume instructions by exploiting that a single Assume instruction can now hold multiple assumptions.

Change-Id: I92048bb46b4d1950ffdae07750fac3d9e760fc1a