Allow R8 markers without a library desugaring identifier when merging
An R8 marker without a library desugaring identifier could come from a
library built with R8 and class file output, so always allow that when
merging.

Right now it is not possible to distinguish this from a similar marker
in DEX. However, merging multiple R8 processed DEX files is an
exceptional case, so not catching that is acceptable with the currrent
marker content.

Bug: 158644894
Bug: 160282413
Change-Id: Ia5c6195694a4f96dfedcfa72c104e1c285afdb51