Account for append() calls in cycles in StringBuilder optimization
This CL introduces a small intraprocedural dataflow analysis to implement the logic for the analysis.

The primary driver of the analysis is IntraproceduralDataflowAnalysis, which will repeatedly apply a given transfer function on the instructions of the basic blocks until a fixpoint is reached (or the transfer function returns a failure, signaling the analysis to stop).

Bug: 163230758
Change-Id: I0d97f67be72cb9933cdd6052a3e4d43e321245b9