Remove ServiceLoader synthetic class double addition
The synthetic class is currently added twice, with this CL it
is now added only once. The line removed prevents the app to
be rebuilt in between the 2 rounds of IR processing since the
synthetic class is added to the appBuilder only after the
second round of IR processing.

This change does not break any test.

Bug: 160939354
Change-Id: I3b3b9de6b9c592602e3c945ca9d5467b24651884