Rewrite regexp retracer to be qualified context aware
Normally a stack trace looks as:
  at foo.bar.Baz.qux(SourceFile:lineNumber)

which is matched somewhat by:
  at %c.%m(%s:%l)

which also means that we normally have the matching groups in sequence
that is similar to what we do for retracing; first the class, then the
method and then the line number.

However, there are no restrictions on the format. Before this CL, we
would not be able to retrace the line-number correctly if the format
was:
  %l %c%m

This is solved here by first having a pass that is finding information
about the string such that everything can be retraced from the first
group. Secondly, we also ensure that we write the string iteratively
instead of patching in all the time.

Finally, this also makes the parser ready for using -verbose

Bug: 159562137
Change-Id: I63f90cd62e03548cb334c8d22868d5988e064f28