Null out SourceFile if attribute is not kept when generating CF
When emitting CF code we should not keep SourceFile if not
specified. If the attribute is specified, specifying sources will not
work. This is also compatible with PG.

For DEX we still maintain the annotation in compat mode since this
will help the most users in case of printed stacktraces.

Bug: 161057618
Change-Id: Ib94207b57f3ee756bb8171f11a609475d68fffb9