Remove method and offset parameters from collectIndexedItems()
The two parameters were used to mutate DexCallSite.method and DexCallSite.offset when compiling for DEX. This is now done in a separate pass in the ApplicationWriter.

This CL additionally removes the definition of collectIndexedItems() on IndexedDexItem. This significantly improves code navigation, since there is no longer dynamic dispatch, but introduces a bit of code duplication in a few places as a result of having to type variables stronger than before.

Another advantage of removing IndexedDexItem.collectIndexedItems() is that there are no more meaningless implementations, such as DexLibraryClass.collectIndexedItems():

  @Override
  public void collectIndexedItems(IndexedItemCollection indexedItems,
      DexMethod method, int instructionOffset) {
    throw new Unreachable();
  }

Change-Id: I5403010c3dc9f6d4eccf58e0dcfa35047061939b