Rewrite EnclosingMethod referring to renamed lambda methods
If a lambda method is renamed as part of lambda desugaring the
EnclosingMethod if any anonymous or local classes of that lambda
method is now rewritten to match the renaming.

This requires that the class with the lambda method and the classes
for the anonymous and local classes of the lambda method are in the
same compilation unit.

Updated the tests to use testForDesugaring and added a tests for R8.

Bug: 158752316
Change-Id: Id41cf812d9aee85fe632f66f3ac5fab764f4b690