Cleanup GraphLens.lookupType()
This CL splits lookupType() into lookupType() and lookupClassType(). As the name suggest, the latter requires a DexType that satisfies isClassType().

This way the handling of array types can be dealt with in a single final method, such that each lens does not need to consider if the given type could be an array type.

Change-Id: I636e8e1adf38543a10623de890359df1c3977b4b