Refine the CF version downgrade due to desugaring
Changed the CF version downgrade to be handled through the CF file
version registered on the classes and methods. When desugaring is
enabled the CF version is downgraded to 1.7 which we try to desugar to.
If the desugaring for some reason cannot be completed the CF version
will be upgraded again.

Bug: 173198228
Bug: 146424042
Change-Id: If468424820ea742d3f9753bd28a3474ee1bdf85f