Provide an option to prefer library desugaring over backporting
This is an intermediate step using the existing configuration file
format for methods on java.util.Objects with additional filtering.

This CL includes tests which should also work when the
configuration file format is changed.

The change to the configuration file is not supposed to be done to
the externally published version.

Bug: 174453232
Bug: 174840626
Change-Id: Ia5f7d931f7693a667a2d79510e6df2eeccb3e919