Allow using non-program contexts that have a rooted program context
This CL splits the definition-for context in the Enqueuer into 'ProgramDefinition rootContext' and 'Definition leafContext'.

This allows providing a non-program context for definition-for, as long that non-program context is derived from a program context.

This CL uses the above to forward the resolved field as a context for recording the type references in the field reference, such that the field definition will be blamed for any missing classes, and not the methods that refer to the field (see Enqueuer#resolveField()).

Change-Id: Ic40680a59cd4402655e39f6bee8f8e317de5be3b