Share logic for analyzing invokes in class inliner
With the new class inliner constraint analysis, there is no longer a need to distinguish methods calls where the instance that is subject to class inlining is being used as the receiver of a virtual invoke, or a non-receiver argument.

This CL therefore unifies the handling of virtual method calls and "extra method calls" in the class inliner.

Additionally, this CL changes the uses of Reason.SIMPLE in the class inliner to Reason.ALWAYS. The motivation for this change is that there is an explicit cost analysis in the ClassInliner, which is responsible for checking if the cost of the inlining does not exceed a given threshold. Therefore, we should not restrict each of the individual called methods to satisfy the 'SIMPLE' predicate.

Bug: 153773246
Change-Id: Ifbe9a918665c32498d0f719e790d989a31a90179