Add generic signature to emulated interfaces
For each emulated interface added to a class the generic signature
will be added as well.

As the emulated interface is added lower in the hierarchy than the
implementation of the original interface the generic arguments from
where the emulated interface is added is evaluated. This is done by
propagating the type arguments up the inheritance hierarchy to where
the original interface was implemented.

Bug: 181750323
Change-Id: I4ba3b8881610b6372473a5c3f0a814245eecf368