Inspect kotlin metadata proto messages before modeling signatures
When asking for an extension signature the kotlin metadata library
synthesizes a signature if possible. When writing the synthesized
signature back it becomes concrete and stored in the message and
string pool and as a result the size increases.

This CL adds inspection of the messages before modeling to ensure we
only model the signatures that are actually needed.

Bug: 185756596
Change-Id: Iff55f02e20cd11a040f7ef75c4a3a15c475827be