Fix verifyRewrittenWithLens() failure from AppServices
This keeps track of which GraphLens was previously used to rewrite the AppServices instance.

This fixes an assertion failure in presence of cycles in the graph lens.

Example:

  Lens 1:
    A -> B
  Lens 2 extends Lens 1:
    B -> A

If AppServices is rewritten with Lens 2, and it references type A (i.e., AppServices is correctly rewritten with Lens 2 from referencing B to referencing A), then the call to verifyRewrittenWithLens() still fails, because A is mapped to B by Lens 1.

Change-Id: Ia05550603b30ebda40e66eee17df0240cb992eef