Change side effect analysis for instance initializers
Previously, instance initializers that did not have any other side effects than initializing the fields on the receiver was marked as not having any side effects.

This CL changes such that we now correctly mark instance initializers as having side effects, when it initializes instance fields on the receiver.

Note that we still have the change that an instance initializer may not have any other side effects than the instance field assignments to the receiver in the InstanceInitializerInfo.

Change-Id: Id909bb7b084e5742ad6ac535958c370ecbd28969