Group equivalent instance initializers in same mergers
This introduces a new (yet to be implemented) InstanceInitializerAnalysis that given an InstanceInitializer returns an InstanceInitializerDescription.

The InstanceInitializerDescription is an abstraction of the instance initializer's code, and is used to group instance initializers into the same mergers, such that they will be mapped to the same constructor in the residual program.

Bug: 189296638
Change-Id: Ib3d804b4ac077b32e9142bd21dc2dc91b9563258