Remove annotations from non-kept items in non-compat mode
RELNOTES: Remove annotations from non-kept items in non-compat mode.

When building with R8 in Proguard compatibility mode, R8 does not remove any annotations from the program that are matched by a `-keepattributes` rule (e.g., the `RuntimeVisibleAnnnotations` attribute). In non-compatibility mode (also called full mode), R8 now removes all annotations from the program, except annotations on classes, fields, and methods that are matched by a `-keep` rule.

For example, to retain all runtime visible annotations for a class `C` it is no longer sufficient to add the rule '-keepattributes RuntimeVisibleAnnotations' in non-compatibility mode. The configuration must also explicitly add a `-keep` rule for class `C`, which can be done with the rule `-keep class C`, or the rule `-keep,allowobfuscation,allowshrinking class C`, which allows both renaming and shrinking of `C`, but still preserves the annotations of `C` if the class is not removed.

Change-Id: Ife273adbcfa1c9a690fc905566b6a87c66353737