Leverage constructor equivalence in final round of class merging
This extends the InstanceInitializerAnalysis to the final round of class merging.

To ensure that the final round of class merging results in a simple renaming, the NoInstanceInitializerMerging policy is updated to check for instance initializer equivalence. Since this equivalence relies on the mapping from the instance fields of source classes to the instance fields on the target class, this must now be built at the point of policy execution.

Change-Id: I8267702df7929db6c0b6dcfa74dd77cca7a33759