Use arraycopy to create enum values array in enum unboxing
This changes such that a shared utility class now contains (n being the largest enum value of any unboxed enum):

  public static final synthetic $VALUES =
      new int[] { 1, 2, 3, ..., n };

Accesses to MyEnum.$VALUES and calls to MyEnum.values() are rewritten to MyEnum$LocalUtility.values() (m being the largest enum value of MyEnum):

  public static final synthetic values() {
    int[] result = new int[m];
    System.arraycopy($VALUES, 0, result, 0, m);
    return result;
  }

This returns a copy of the underlying $VALUES array when accessing it via a field read. The $VALUES array generally only has a single use, which is followed by a call to clone(). This CL therefore identifies the call to clone() and removes it.

As a result of this, MyEnum.$VALUES == MyEnum.$VALUES will return false (but Arrays.equals(MyEnum$VALUES, MyEnum$VALUES) is true). If this turns out to be a problem, we can consider disabling enum unboxing of enums that have unsupported reads of $VALUES.

Change-Id: I543fb1b73bf34faa12486d00137ca1d0d10cf863