Update configuration of library desugaring testing
Add a builder for constructing the library desugaring test configuration.
This allows better control of the configuration used, and have the added
option of building the desugared library and adding it to the runtime
classpath automatically, as this is almost always desired.

This now requires that the library for the compilation is explicitly set
to one that has the APIs used by the test. The library for P has been
used, where this was missing.

The existing uses of the enableCoreLibraryDesugaring is not changed, and
will still have to explicitly build and add the desugared library to the
runtime classpath. This is done by having an option in the builder for
the desugared library test configuration to say to not add the desugared
library to the runtime classpath.

This is in preparation for easier testing different configurations.

Bug: 158543446
Change-Id: Ia88052008e7c74210cd865e2483446ba6fafa554