Classify methods as checkNotNull() methods for enum unboxing
This adds a new EnumUnboxerMethodClassification to the MethodOptimizationInfo.

A given method is classified as a checkNotNull() method by the enum unboxer if it is a static method with a parameter of type java.lang.Object, which has a single if-zero user.

Follow up work will use this piece of optimization info to allow unboxing of enums with checkNotNull() users.

Bug: 192037990
Change-Id: I9024de37c46da742be63ac8d412fde458206e488