Update CF to CF desugaring to have a prepare step for adding methods
In the prepare step each desugaring can synthesize methods on
program classes so that these methods are present when the actual
desugaring take place.

This does not affect true synthetics, only synthetic methods which
must be synthesized on program classes.

Also the D8 addition of bridges for access from methods on classpath
is not affected.

Bug: 192310793
Change-Id: If645f0e3446fb5fb757ab7a7437c2497ceb2f6b9