Ensure that all methods and fields are assigned an api reference level
This CL will ensure that all members will have an API-level defined. If
api modeling is enabled, min-api will be used as the base for computing
otherwise UNKNOWN will be used.

This works because we do allow inlining and merging of classes where
both are references are UNKNOWN. In principle, we could disallow this
but for now this is the design.

Bug: 188388130
Change-Id: I33298168dd511af74803d8f5360ff9630b9b31c9