Update isPinned() to return true with -dontoptimize
This is a step towards removing the `noShrinking` and `dependentNoShrinking` collections.

In the KeepInfo, we only model the modifiers that are not overruled by the global configuration. For example, in case of -dontshrink, we don't set the KeepInfo#allowShrinking to false for the items matched by -keep rules, since this already follows from InternalOptions.isShrinking(). For this to work, we need `isPinned()` to consult the `GlobalKeepInfoConfiguration`.

Bug: 192924387
Bug: 191617670
Change-Id: I89673f8ea7bb6eb0cea4ebf1d20ab7049f6df2fe