Fix desugared library wrapper concurrency issues
Fix the red bots

There were various concurrency issues because we used to:
(1) create the wrapper class
(2) introduce the wrapper virtual methods
(3) introduce the conversion method
Once (1) is done, the wrapper class could be accessed for
concurrent look-ups, (2) and (3) could happen concurrently.
The fix was not easy due to high amount of recursion.

To simplify the process, in this CL:
- classpath wrappers are created in one step, then made accessible.
- program wrappers are generated in 2 steps before instruction
processing,(1) as empty shells with the conversion method,(2)
the virtual methods are added assuming the conversion methods are
present.

Change-Id: I167aa1f14ac7995001cc45e3730132f078f5a98b