Widen parameter states to unknown when dynamic type is trivial
This passes the static type of a given parameter to the join methods in ParameterState. This allows the join method in BottomClassTypeParameterState and ConcreteClassTypeParameterState to check if the resulting computed dynamic type is trivial according to the static type, in which case the state can be replaced by 'unknown' (if the abstract value is also unknown).

Change-Id: I927cdc06a281b5680128875e134a4e130d8eec00