Drop type bounds for arguments to non-overriding/overridden virtual methods
This computes the set of virtual methods that do not override other methods and are not themselves overridden by other methods.

Such methods can be treated as direct methods, where argument information is stored in a monomorphic method state (in which argument information is not qualified by dynamic type bounds).

This leads to less memory overhead and reduces the amount of propagation in the post processing phase.

Change-Id: Ia8f00b00b21c0bd0fbe6a73271dbe3d255e4a91f