Ensure correct lens rewriting of reprocessing sets
This fixes some subtle issues that may arise due to inconsistent rewriting of reprocessing sets in the enum unboxer, the inliner, and the staticizer.

We now consistently use a LongLivedProgramMethodSet for the reprocessing sets, and maintain the invariant that all methods in these sets are always rewritten up until the same graph lens.

This ensures that we never reapply parts of a graph lens to the same method, which may lead to incorrect rewriting.

Change-Id: I6ce3fd3d52e36f3090901c82805dad8806a99def