Don't use accessors for lambdas to interface companion methods.
This CL adds in nested desugaring so that the invoke-dynamic target is
desugared *in context* of the invoke-dynamic.  This allows desugaring
the interface targets in the correct context and avoids the need to
introduce access bridges.

This should allow moving the finalization of interface companion methods
up into method processing in D8.

Change-Id: I6bd011f8a88770d1b27a246b9c32d9b0eae55b14