User LinearFlowIterator in outliner to outline inline positions
Consider the following code:

public static void foo() {
  inline();
  bar();
}

public static void inlinee() {
  baz();
}

public static void bar() {
  ...
}

public static void baz() {
  ...
}

When we inline inlinee into foo we build we add a simple goto block for
the instruction to invoke baz(), which is perfectly fine to outline.

This CL changes the view of an outline from a single block to a linear
flow possibly spanning multiple blocks to allow inlined instructions to
be included in outlines.


Bug: 204749490
Change-Id: Ibfaa5931beee63e38485de163970bcf6d249e7d0