Emit a pc-to-pc line mapping in debug info tables.
The use of pc-to-pc line mappings allows most methods to share the the
same debug_info_item for the encoding of their debug information. The
shared info items reduces the size of the DEX files and unifies the line
mappings such that they are the same on legacy devices as well as on
newer devices that support emitting the PC directly.

Bug: 205910335
Change-Id: I557f8c47f481536c4a8470aafc18bd9ec7082e54