Rewrite all users of dynamic types to use DynamicType
This refactors the optimizations that use dynamic upper bound and/or lower bound types into using DynamicType.

Since the dynamic types are used to propagate not null information, this also introduces a new NotNullDynamicType. This has two advantages:

(i) All dynamic types that are equal to their corresponding static type, except that the dynamic type guarantees non nullness, can be mapped to the new NotNullDynamicType. Since NotNullDynamicType is a singleton this leads to fewer instances of DynamicType.

(ii) Dynamic types are generally not available in D8. The NotNullDynamicType instance can be used, however, to model that certain library methods are guaranteed to return non null even in D8.

Change-Id: I0033796cf070e35c05529a8c5fab3e85f1097c71