Light weight call graph construction for multi caller inlining
This CL generalizes the existing call graph construction in a way that allows different call graphs using the same underlying call graph construction algorithm.

To this end, a new MultiCallerInlinerCallGraph is introduced where nodes are instances of MultiCallerInlinerNode. Unlike the existing CallGraph nodes, these nodes only store the number of callers (and not also the caller, callee, reader, and writer sets).

Change-Id: I7afb735485f92fd1346f139be81f403ad47d8a4a