Reproduce CONSTANT_Dynamic desugaring issue with getDeclaredMethods
The first argument of a CONSTANT_Dynamic bootstrap method has
parameter type java.lang.invoke.MethodHandles.Lookup. That type
is not defined on Android with API levels below 26. Desugaring
of CONSTANT_Dynamic leaves the bootstrap method in the program
causing getDeclaredMethods on a class which had a CONSTANT_Dynamic
bootstrap method to fail with ClassNotFoundException.

Bug: 210485236
Change-Id: I0527ed10d1255a3f23da99f847f75bc14313b7ba