Don't erase dynamic argument types not needed for reprocessing
If the dynamic type of a given parameter in the program was deemed to not have any effect on reprocessing, we removed the dynamic type information.

This CL changes the argument propagator to not prune such dynamic type information (as it is needed for strengthening parameter types), meanwhile making sure methods are not reprocessed due to the newly retained dynamic types.

Change-Id: Iacdb66d0cb5b56958982c1969047a364b9783067