Remove unused arguments in presence of instance initializer collisions
Previously unused argument removal would bail out in presence of instance initializer collisions, e.g.,

  void <init>()
  void <init>(Unused unused)

To extend unused argument removal to "effectively" unused arguments, it is important that all unused arguments are removed.

Example: If the only uses of the parameter 'effectivelyUnused' in method foo(Unused) is a call to <init>(effectivelyUnused), then the parameter 'effectivelyUnused' can be removed from foo(Unused). However, this only works if the parameter 'unused' is removed from <init>(Unused).

  void <init>()
  void <init>(Unused unused)
  void foo(Unused effectivelyUnused)

To this end, this CL ensures that the unused argument from <init>(Unused), at the cost of appending a fresh argument to the instance initializer.

Bug: 214901256
Change-Id: Ia357f1f303537c2e10064032a595169c2a018c3a