Revert "Revert "Add implementation of hashCode and equals to desugared library wrappers""
This reverts commit 4765beff28beecadfb0eeaa0fe45de276724a24f.

Reason for revert: Fixed compilation

Add implementation of hashCode and equals to desugared library wrappers

The implementation of hashCode just forwards to the wrapped instance.

The implementation of equals checkes for a wrapped argument and if
found passes the the wrapped instance when forwarding the call to
the wrapped instance.

Bug: b/230800107
Change-Id: Iea18f0d7ea426dbc6beeffcb6420d14fc6a6cf1d