Avoid lossy joins in cf analysis
This changes InitializedNonNullReferenceFrameType into being the base type of InitializedNonNullReferenceFrameTypeWithInterfaces and InitializedNonNullReferenceFrameTypeWithoutInterfaces.

The class InitializedNonNullReferenceFrameTypeWithoutInterfaces is backed by a DexType, similar to the existing InitializedNonNullReferenceFrameType.

The newly introduced InitializedNonNullReferenceFrameTypeWithInterfaces is backed by a ReferenceTypeElement. Instances of this are (only) used when computing the join of two reference types in the CF analysis to avoid loss of information from the joins. Thus instances of this class will not appear in any CfFrame in CfCode.

Fixes: b/236581210
Change-Id: I56bba15e1821e8d36a8757380b90eb3688305acd