Change handling of enums in annotations when generating class files
This is a follow-up to the previous change "Fix handling of enums in
annotations when generating class files". With this change the enum
fields references in annotations are no longer renamed. This ensures
that an R8 processed class file using enum annotations can be correctly
be processed a second time with R8 - also when compiling to DEX.

Fixes: b/236691999
Change-Id: I6cab30f8798d26e2ac8016c68da48a25b0337d61